fix(externals): handle absent storage properly
The save/load methods already contain the logic that
getstorage/putstorage methods that return nothing indicate that the
method did not work for some reason, and the procedures should keep on
looking for other suitable methods.

It's safe to assume that when a dedicated getstorage/putstorage method
is not availible, that the save/load procedure should behave like this,
so it makes more sense for the fallback methods to return nothing than
raise an error.