tweak(externals): ensure write iostreams closed
Some methods (like CSV.write) when applied to an IO stream write but do
not close or flush the IO stream. However, they do return the IO stream.
So, to enable more convenient write functions we can just check the
result and if it's an IO steam, ensure that it's closed.