feat(externals): de-duplicate supportedtypes
Multiple definitions with different from types can provide multiple of
the same result type. Furthermore, it appears that the type order
corresponds to the order of method definition, from most recently
defined to oldest. I'm inclined to think that most often the
expected/preferred type methods are defined first, as so it's probably
best to reverse the list of types for the sake of most cases.