fix(parser): be more lax with spec structure
This way a data set written solely in terms of the loader can be parsed
without causing an error. E.g.

[[one.loader]]
driver="raw"
value="1"

Before this would result in a single Dict instead of a vector of Dicts,
and so iteration would produce pairs, and this would cause a method call
error with the DataSet constructor.