fix: rename REPL extension to avoid name conflict
It seems that something funky can happen with package extensions (of
different packages) with the same name. It seems like we want a globally
unique name here, or a bugfix in Pkg. Given this (buggy) behaviour is
already part of two minor Julia releases (and soon a third), let's just
use a globally unique package extension name.

    julia> using Pkg
    ERROR: LoadError: ArgumentError: Package REPLExt does not have DataToolkitBase in its dependencies